stlport compile warnings fixed compile fix for axemod
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@3394 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/DayWatcherThread.cpp
@@ -273,8 +273,8 @@ void DayWatcherThread::update_arena()
 				}
 			}
 
-			arenapointsPerTeam[0] = max(arenapointsPerTeam[0],arenapointsPerTeam[1]);
-			arenapoints += max(arenapointsPerTeam[0],arenapointsPerTeam[2]);
+			arenapointsPerTeam[0] = (uint32)max(arenapointsPerTeam[0],arenapointsPerTeam[1]);
+			arenapoints += (uint32)max(arenapointsPerTeam[0],arenapointsPerTeam[2]);
 
 			if(orig_arenapoints != arenapoints)
 			{
